# week3_lita/openapi.yaml
openapi: 3.0.3
info:
  title: Neon Database CRUD API
  description: API Richardson Maturity Model Level 3 dengan navigasi HATEOAS untuk data siswa, guru, dan staff.
  version: 1.0.0
servers:
  - url: /
    description: Current server (local or deployed)
paths:
...
  /api/students:
    get:
      summary: List all students
      responses:
        '200':
          description: Student resources with hypermedia links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonResource'
    post:
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonInput'
      responses:
        '201':
          description: Student created with hypermedia links
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResource'
...
  /api/teachers:
    get:
      summary: List all teachers
      responses:
        '200':
          description: Teacher resources with hypermedia links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonResource'
    post:
      summary: Create a teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonInput'
      responses:
        '201':
          description: Teacher created with hypermedia links
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResource'
...
  /api/staff:
    get:
      summary: List all staff
      responses:
        '200':
          description: Staff resources with hypermedia links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonResource'
    post:
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonInput'
      responses:
        '201':
          description: Staff created with hypermedia links
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResource'
...
components:
...
  schemas:
    PersonInput:
      type: object
      required: [nama, umur, tanggal]
      properties:
        nama:
          type: string
          example: Lita
        umur:
          type: integer
          example: 20
        tanggal:
          type: string
          format: date
          example: '2006-05-12'
    Person:
      type: object
      required: [id, nama, umur, tanggal]
      properties:
        id:
          type: integer
          example: 1
        nama:
          type: string
          example: Lita
        umur:
          type: integer
          example: 20
        tanggal:
          type: string
          format: date
          example: '2006-05-12'
    Link:
      type: object
      required: [href, method]
      properties:
        href:
          type: string
          example: /api/students/1
        method:
          type: string
          enum: [GET, POST, PUT, DELETE]
          example: GET
    PersonResource:
      allOf:
        - $ref: '#/components/schemas/Person'
        - type: object
          required: [_links]
          properties:
            _links:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/Link'